build,storage,systemtests: rename 'null' driver to 'test' driver
