+++ systemtests/testdata/test/global1.json
@@ -2,5 +2,5 @@
   "TTL": 30,
   "Debug": true,
   "Timeout": 10,
-  "Backend" : "null"
+  "Backend" : "test"
 }
